requirements.txt

einops
glfw
imageio
imgaug
imutils
jupyter
matplotlib
mmcv==1.3.8
numpy==1.23.1
open3d
opencv-python
pandas
pydantic-cli
pydantic-yaml
PyOpenGL
PyOpenGL_accelerate
pyyaml
scipy
tensorboard
timm==0.4.12
torch==2.0.1
torchvision==0.15.2
torchaudio
tqdm

